Add files via upload
This is the database, version 1 for the hotel reservation system. At the moment this database is essentially set up for the use by the hotel employee, ie receptionist or managers etc to set up the customers details for their rooms, payment and loyalty points program. A second database or a new table would be created to house customer login to book their room etc